Rename task_helpers to tasks
This fixes a problem where celery is unable to autodiscover tasks in
the task_helpers module when DEBUG setting is False.

Test by toggling your DEBUG setting between True and False and do the
following each time:
1. Restart celery `sudo service celeryd restart`
2. Inspect registered tasks `envdir /etc/mmw.d/env celery -A
'mmw.celery:app' inspect registered`

The registered tasks should match whether or not Django is in DEBUG
mode.

Connects #419
